+++ Changelog.md
@@ -1,3 +1,8 @@
+### v0.2
+* migrate to bevy 0.17
+* made feature "perf_ui" optional in showroom example
+* removed iyes_perf_ui dependency (and use the bevy FpsOverlayPlugin instead)
+
 ### v0.1.2
 * Made `generate_tree_meshes()` public, to be used without the `TreeProceduralGenerationPlugin`